+++ scraper/test-new-sources.ts
@@ -25,6 +25,7 @@ import { scrapeWorkable } from './engines/workable';
 import { scrapeAvanti } from './engines/avanti';
 import { scrapeVipCloud } from './engines/vipcloud';
 import { scrapeWorkzoom } from './engines/workzoom';
+import { scrapeSapWebDynpro } from './engines/sap-webdynpro';
 import { scrapeBrassRing, scrapeEdmontonPhenom, scrapeHaltonHills, scrapeNipissing, scrapeNorthernCollege, scrapePickering, scrapeStClairCollege, scrapeStLawrenceCollege, scrapeVaughanPL } from './engines/custom';
 
 const REQUIRED_SUCCESSFUL_RUNS = 3;
@@ -133,6 +134,8 @@ const SOURCES = {
     scrapeVipCloud(db, context, 'https://georgina-careers.vipcloud.ca/default', 'Town of Georgina'),
   'County of Renfrew': (db: Client, context: BrowserContext) =>
     scrapeWorkzoom(db, context, 'https://curos.ca/curos/COR2302/V/TRBJO_PUBLIC', 'County of Renfrew'),
+  'Manitoba Hydro': (db: Client, context: BrowserContext) =>
+    scrapeSapWebDynpro(db, context, 'https://careers.hydro.mb.ca/sap/bc/webdynpro/sap/hrrcf_a_unreg_job_search', 'Manitoba Hydro'),
   'Government of Alberta': (db: Client, context: BrowserContext) =>
     scrapeJobs2Web(db, context, 'https://jobpostings.alberta.ca/go/All-Jobs-GoA/2617217/', 'Government of Alberta'),
   'Ontario Energy Board': (db: Client, context: BrowserContext) =>